LayoutTests/imported/w3c/web-platform-tests/custom-elements/registries/element-mutation-expected.txt


PASS An element with global registry should not change its registry when moved into a shadow tree with scoped registry.
FAIL An element with scoped registry should not change its registry when moved out of the shadow tree. assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
FAIL An element with scoped registry should not change its registry when moved into another shadow tree with different scoped registry. assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
PASS Declarative shadow DOM with shadowrootcustomelementregistry attribute without registry initialized should remain null registry after adoption.
